cryptodev: Move to 1.12 revision
Cryptodev out of kernel module and suite has yet another set of
challenge with broken API compatibility. However, the complete list of
patches beteween dunfell's default 1.10 and current latest 1.12 revisions:
 e0c25e289d6b (tag: cryptodev-linux-1.12) Release version 1.12
 55c6315058fc Fix build for Linux 5.11-rc1
 2f5e08aebf92 Fix build for Linux 5.9-rc1
 e8263d2d682b remove VLA usage from authenc.c
 dafa42cadc86 Fix for change in Kernel 4.4.168
 f4d3653c8dcb Release version 1.11
 9e765068582a Fix build for Linux 5.8-rc1
 7e72f6758f2d enabled the support for TLS1.1 - AES128-SHA1 - AES256-SHA1
 98b163a23f6b cryptlib.c: fix build on kernel v5.5+
 a1e738aef624 tests: cipher-aead-srtp: Fix cao->dst alignment
 f971e0cd4a0e Fix module loading with Linux v5.0-rc5

We have the following choices
1) just cherry-pick minimum set of kernel patches relevant for kernel
compatibility
OR
2) switch to a newer version with richer set of fixups

Attempting to switch to 1.12 version with minimal changes as a better
option to get a wider swath of kernel compatibility (v5.5 to v5.11), vs
the default of v5.4 kernel in dunfell vanilla.

Signed-off-by: Nishanth Menon <nm@ti.com>