init-manager-*: Leverage oe-core method of choosing init manager
Change arago over to the oe-core init-manager setup for all images instead
of a mixture of systemd/sysvinit based on perceived image size.  With the
migration to a more targeted initramfs we have the space in the image to
support systemd and not require sysvinit for the smaller images.

We will still support setups that use ARAGO_SYSVINIT, but the preferred
method will be to set INIT_MANAGER in the local.conf with the default
being systemd.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>