tisdk-tiny-image: Add dependency on tisdk-tiny-initramfs
Even if we do not tell the build that the initramfs should be bundled
into the kernel, we want to build the initramfs with every build so that
it can be manually specified through alternative testing schemes.

Also, we no longer want to build the cpio from the tiny-image, that is
the purpose of the dedicated tisdk-tiny-initramfs.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>