meta-arago-distro: add meta-clang dependency and llvmpipe
Devices that must use software rendering should use llvmpipe over the
default softpipe. There are significant performance improvements
associated with this change for devices like AM62A which lack a GPU.

This, unfortunately is only available on armv6+ devices so we'll
conditionally drop this for the omapl138 device. (I'd switch this
directly based on platform architecture but this doesn't appear to work
correctly for PACKAGECONFIG strings in the distro.conf)

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>