packagegroups: crypto: Add pkcs11 tools and example TAs
The OP-TEE package contains some loadable TAs including one used for
pkcs11, add this package to the base image.

To make use of the PKCS#11 API (Cryptoki) that this TA provides, one
can use libckteec already included with optee-client. An example
app that can make use of that library is pkcs11-tool which is part of
opensc. Add this package also.

For example, try:

$ pkcs11-tool --show-info --module /usr/lib/libckteec.so.0

Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>