tisdk-sw-manifest: Switch up the TI software manifest creation
We have been seeing an odd pseudo bug related to how the TI software
manifest was being created.  Prior to this change we needed to run the
SDK installer in order to get some of the packages that needed to be
included in the manifest.  But installing the toolchain would randomly
cause a pseudo abort error.

This change has the part of the manifest that is needed from the
toolchains to be created at the same time as the toolchain.  Then the
software manifest can just include those sections without needing to run
the installer.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>