packagegroups: base-tisdk: Remove iperf from packagegroup
We want iperf3 which is already in this list, iperf is the older
v2.0 that is only needed by some old test scripts. It is already
included in the test packagegroups.

Without this patch we cannot build base or thinlinux images without
the meta-arago-test layer. That layer is not listed in the LAYERDEPENDS
and so should be considered optional for meta-arago-distro images.

Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>