tisdk-bundle: bring this class on par with current image.bbclass
As tisdk-bundle was based on the image class long time ago, it inherited
some of its old attributes and artifacts used back in the days. There
were several updates and reworks done to this class through the years to
keep it up to date and in sync with upstream image class, but some pieces
fell through the cracks. In recent and not so recent years upstream image
class have deprecated some things, changed and extended others, such as
consolidated umask handling, improved sstate handling, expanded list of
cleaned directories and changed stamping and locking policy.

Since tisdk-bundle inherits image class, corresponding attributes for
do_rootfs, do_image and do_image_complete don't need to be modified here
and can be left to their default values, as well as the order of those
tasks.

This change addresses the issue with sstate/deploy file collisions we've
been seeing at times lately resulting in pseudo abort errors. And it
improves long term maintainability of tisdk-bundle class as well.

Signed-off-by: Denys Dmytriyenko <denys@konsulko.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>