qtbase: Patch to support xkbcommon 1.6.0
xkbcommon 1.6.0 (and newer) remove 4 defintions that qtbase seems to use.
Backport an upstream patch to look to see if they are defined and
ignore them if they're not.

This is a temporary patch to fix our master builds while we wait for the
meta-q5 patch [1] to make it to the master branch.  This will be
reverted once that patch is promoted.

[1] https://github.com/meta-qt5/meta-qt5/pull/545

Signed-off-by: Ryan Eatmon <reatmon@ti.com>