receipes-multimedia: gstreamer: Add patches for AM62P
Add patches to prefer contiguous video format and supporting
dmabuf-import mode for video decoder.

This is to support use-cases involving video codec with other components
viz. GPU, Camera e.t.c.

For 0002-v4l2-Give-preference-to-contiguous-format-if-support.patch :
There are certain blocks such as camera, GPU which only seem
to support well the contigous formats (sometimes at the S/W level)
which doesn't go well with gstreamer v4l2 which prefers non-contigous
by default and we fail at negotiation.
The crux of the issue is that unlike v4l2 gstreamer has only single GST
video format for both NV12 and NV12M (non-contigous) and it prefers to
advertise latter during caps negotiation.
There was a proposal made long back to have separate GST video
formats for NV12 and NV12M but that didn't converge, perhaps they
wanted to support this with a VideoMeta based scheme as per
logs [0].

For 0001-v4l2-Changes-for-DMA-Buf-import-j721s2.patch :
This is a local patch which complements a patch already made
available in driver to support dmabuf import for the decoder.
Here the gist of the issue is that v4l2 framework expects userspace
to make sure it uses same v4l2 buffer index each time with same backing
memory as kernel framework relies on these indices to identify frame
ordering [2].

There again is discussion around this to support this in upstream but
the upstream patch is not yet merged [1].

[0]:
https://gitlab.freedesktop.org/gstreamer/gst-plugins-good/-/issues/414
[1]:
https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/4550/diffs?commit_id=b12eaf65f489bab50917e22242758e000d69df58
[2]:
https://www.kernel.org/doc/html/latest/userspace-api/media/v4l/dev-decoder.html

Signed-off-by: Devarsh Thakkar <devarsht@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>