arago.conf: remove x11 again
We don't support x11, and while it can currently be enabled in tandem
with our graphics stack right now without explicitly causing issues on
it's own, Yocto still treats the presence of x11 as a distro feature to
mean that desktop GL is present.

This assumption leads to a lot of incorrectly configured packages in our
case so to prevent future incompatibility with layers adjusting their
PACKAGECONFIG values with this logic we should explicitly remove x11
again.

This decision will likely flip again if oe-core introduces a standard
"opengles" feature instead of just the blanket "opengl" one they
currently use.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>