meta-arago: distro: add a bbappend to patch the chromium browser
This applies a patch for meta-browser/chromium-ozone-wayland to make the
sandboxing work with the Imagination GPU components without the need for
the --no-sandbox flag.

GPU acceleration in Chromium is dependant on IMG DDK 23.3.  It works
across AXE/BXS/8XE GPU's.   No acceleration is expected for SGX.

Also add an upstream patch to stop Chromium from segfaulting when it is
run with no input devices connected to the board.
Patch was retrieved from:
https://chromium.googlesource.com/chromium/src/+/323077958301bc321d840a2c2b983ab469934753

Signed-off-by: Darren Etheridge <detheridge@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>