Add description and summary for jailhouse image recipe
Image recipe for jailhouse does not have a summary and description. Due to
which oe-layersetup also shows No Summary available for tisdk-jailhouse-image
as compared to other targets for which we have proper summary.

So to ensure good practice and help other users understand the purpose and
functionality of recipe, add a summary and description for jailhouse image.

Signed-off-by: Paresh Bhagat <p-bhagat@ti.com>
Acked-by: Chirag Shilwant <c-shilwant@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>