conf/images: Add ARAGO_IMAGE_SUFFIX to image names
When attempting to merge the build artifacts from bitbake into a new release
tarball coupled with running multiple bitbake builds for non-RT and RT kernels,
the common naming for the image files poses an issue that requires a manual rename
of the RT image file to make it unique.

This patch adds a new ARAGO_IMAGE_SUFFIX variable that is normally "", but for RT
builds is set to "-rt" so that the image files are different.

Backport from master: 9e10acff265c4dfcd150329641a71a6cd55c6572

Signed-off-by: Ryan Eatmon <reatmon@ti.com>