all: purge matrix ui and all dependent packages
Purge the old Matrix UI. This thing was cool but it required modifying
too many other packages, does not have a dedicated maintainer, and
relied on deprecated components.

There have been 2 replacements for this tool that have cropped up in the
last year alone. Surely one of those is suitable for the task now. If
not, now there is proper motivation.

Signed-off-by: Randolph Sapp <rs@ti.com>
Acked-by: Andrew Davis <afd@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>