branding-ltsprep: enable newer pvr drivers
The newer drivers require a dt binding change for platforms that is only
present in the ltsprep kernel. Enable them with the new kernel here.

Also pull in the newer pvr-mesa package here as well. It's not strictly
required for this driver version but it is recommended. It still needs
some attention with regards to SGX so it will not be enabled by default
right now.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>