meta-arago-distro: chromium-browser: add support for Chromium v123
Add support for Chromium v123 and remove the v111 support as this
is not functional on Scarthgap.

This fixes up the wrapper script to use --use-gl=angle instead of
--use-gl=egl when launching Chromium.  With the PowerVR drivers
everything works better if we go through ANGLE with never versions of
Chromium.

Also apply a couple of patches (these are not strictly needed) the first
reduces a permissions warning that gets logged to stderr,  however it is
not fatal because chromium falls back to using another allocation method
which succeeds.   The second is an upstream patch that fixes an issue
where the buffer object size that is getting allocated is incorrect.
It hasn't been shown to be a problem in our testing, but just in the
abundance of caution this patch is going to be included.

Signed-off-by: Darren Etheridge <detheridge@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>