vulkan-cts: add workaround for createMeshShaderMiscTestsEXT
Remove the CXX flag "-O2" for GCC 13 and 14 on aarch64 platforms.
There's a bug with ARM GCC that breaks the iteration of "types" in the
createMeshShaderMiscTestsEXT function. This issue is not present for
clang or x86_64 GCC 14.

It seems that the array is not initialized before the first iteration.
In testing this can result in a random value being used. This can
manifest in LINES type being processed twice, resulting in the following
error:

FATAL ERROR: Failed to initialize dEQP: Test case with non-unique name
'no_lines' added to group 'misc'.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>