recipes-core: images: tisdk-xen-domu: init
Arago TI SDK image meant to be run as rootfs for Xen DomU for
demonstration purposes and ease of use.  This image is derived
from tisdk-tiny-initramfs image.

This patch also sets the XEN_RAMFS_IMAGE variable which is set to the
image name for the ease of use by other recipes that plan to use it as a
cpio archive or for other purposes. meta-tisdk uses the variable to
install the said cpio archive under tisdk-xen-image.

Signed-off-by: Amneesh Singh <a-singh21@ti.com>